(feat) Order's price as big integer

+++ client/src/main/kotlin/com/rarible/looksrare/client/model/v1/LooksRareOrder.kt
@@ -29,7 +29,7 @@ data class LooksRareOrder(
 
     val amount: Int,
 
-    val price: BigDecimal,
+    val price: BigInteger,
 
     @JsonFormat(shape=JsonFormat.Shape.STRING, pattern="s")
     val startTime: Instant,
